Use dpkg-divert in gnome keyring daemon wrapper setup.

+++ misc/gnome-keyring-disable-ssh-agent.sh
@@ -8,6 +8,9 @@
 # This is because it is otherwise overly cumbersome to disable its ssh-agent on some platforms.
 # This is based on this post http://askubuntu.com/a/412794 by JanKanis.
 
+dpkg-divert --divert /usr/bin/gnome-keyring-daemon-wrapped --rename /usr/bin/gnome-keyring-daemon \
+	|| dpkg-divert --divert /usr/bin/gnome-keyring-daemon-wrapped /usr/bin/gnome-keyring-daemon
+
 if [ "`readlink -f "/usr/bin/gnome-keyring-daemon"`" == "/usr/bin/gnome-keyring-daemon-wrapper" ]; then
 	echo "Wrapper seems to be already installed, aborting" >&2
 	exit 1
@@ -23,7 +26,9 @@ if [ -L "/usr/bin/gnome-keyring-daemon" ]; then
 	exit 1
 fi
 
-mv /usr/bin/gnome-keyring-daemon /usr/bin/gnome-keyring-daemon-wrapped
+if [ -f "/usr/bin/gnome-keyring-daemon" ]; then
+	mv /usr/bin/gnome-keyring-daemon /usr/bin/gnome-keyring-daemon-wrapped
+fi
 echo '#!/bin/sh' > /usr/bin/gnome-keyring-daemon-wrapper
 echo 'exec /usr/bin/gnome-keyring-daemon-wrapped --components=pkcs11,secrets,gpg "$@"' >> /usr/bin/gnome-keyring-daemon-wrapper
 chmod +x /usr/bin/gnome-keyring-daemon-wrapper
